Account for true LBA48 sector count reporting in IDE driver. Cleanup fatx.h Renamed BackupPartTbl struct to BackupPartTable in ff.c Added more stuff to FatFSAccessor

+++ drivers/ide/BootIde.c
@@ -1530,7 +1530,7 @@ int BootIdeFlushCache(int nDriveIndex)
     return result;
 }
 
-int BootIdeGetSectorCount(int nDriveIndex)
+unsigned long long BootIdeGetSectorCount(int nDriveIndex)
 {
     return tsaHarddiskInfo[nDriveIndex].m_dwCountSectorsTotal;
 }
