+++ drivers/ide/BootIde.h
@@ -89,7 +89,7 @@ int BootIdeWriteSector(int nDriveIndex, void * pbBuffer, unsigned int block, uns
 int BootIdeWriteMultiple(int nDriveIndex, void * pbBuffer, unsigned int startLBA, unsigned short len, unsigned char retry);
 
 int BootIdeFlushCache(int nDriveIndex);
-int BootIdeGetSectorCount(int nDriveIndex);
+unsigned long long BootIdeGetSectorCount(int nDriveIndex);
 int BootIdeGetSectorSize(int nDriveIndex);
 
 bool driveToggleSMARTFeature(int nDriveIndex, unsigned short smart_cmd);
